Move `ChunkKeySeparator` to `metadata` namespace
