+++ src/metadata/array/chunk_key_separator.rs
