+++ src/metadata/v2.rs
@@ -32,14 +32,14 @@ mod tests {
 
     use super::*;
     use crate::{
-        array::{chunk_key_encoding::ChunkKeySeparator, DataType},
+        array::DataType,
         metadata::{
             array_metadata_v2_to_v3,
             v3::codec::{
                 blosc::{self, BloscCodecConfigurationV1},
                 transpose::{self, TransposeCodecConfigurationV1},
             },
-            ChunkShape, Endianness,
+            ChunkKeySeparator, ChunkShape, Endianness,
         },
     };
