+++ src/metadata/v3/chunk_key_encoding/default.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 use derive_more::Display;
 
-use crate::metadata::v3::ChunkKeySeparator;
+use crate::metadata::ChunkKeySeparator;
 
 /// The identifier for the `default` chunk key encoding.
 pub const IDENTIFIER: &str = "default";
